requirements.txt
google-cloud-aiplatform==1.74.0
msal==1.31.1
psycopg2==2.9.10
python-dotenv==1.0.1
streamlit==1.41.1